Handling of exceptions

+++ db/db_user.py
@@ -2,6 +2,7 @@ from sqlalchemy.orm.session import Session
 from schemas import UserBase
 from db.models import DBUser
 from db.hash import Hash
+from fastapi import HTTPException,status
 
 def create_user(db:Session, request: UserBase):
     new_user = DBUser (
@@ -22,10 +23,21 @@ def get_all_users(db:Session):
 
 
 def get_user(db:Session, id):
-    return db.query(DBUser).filter(DBUser.id == id).first()
+    user = db.query(DBUser).filter(DBUser.id == id).first()
+
+    if not user:
+        raise HTTPException(status_code = status.HTTP_404_NOT_FOUND,
+        detail=f"user with id: {id} not found")
+    return user
 
 def update_user(db:Session, id:int, request:UserBase):
     user = db.query(DBUser).filter(DBUser.id == id)
+
+    if not user.first():
+        raise HTTPException(status_code = status.HTTP_404_NOT_FOUND,
+        detail=f"user with id: {id} not found")
+    return user
+
     # Handle any exceptions
 
     user.update({
@@ -40,6 +52,10 @@ def update_user(db:Session, id:int, request:UserBase):
 def delete_user(db:Session, id:int):
     user = db.query(DBUser).filter(DBUser.id == id).first()
     # Handle any exceptions
+    if not user:
+        raise HTTPException(status_code = status.HTTP_404_NOT_FOUND,
+        detail=f"user with id: {id} not found")
+
     db.delete(user)
     db.commit()
